chore: make sure we are actually installing packages locally
