+++ packages/google-auth-library-nodejs/package.json
@@ -23,9 +23,9 @@
   "dependencies": {
     "base64-js": "^1.3.0",
     "ecdsa-sig-formatter": "^1.0.11",
-    "gaxios": "7.1.3",
-    "gcp-metadata": "8.1.2",
-    "google-logging-utils": "1.1.3",
+    "gaxios": "file:../gaxios",
+    "gcp-metadata": "file:../gcp-metadata",
+    "google-logging-utils": "file:../logging-utils",
     "gtoken": "^8.0.0",
     "jws": "^4.0.0"
   },
