+++ packages/logging-utils/package.json
@@ -3,6 +3,7 @@
   "version": "1.1.3",
   "description": "A debug logger package for other Google libraries",
   "main": "build/src/index.js",
+  "types": "build/src/index.d.ts",
   "files": [
     "build/src"
   ],
