+++ packages/logging-utils/src/index.ts
@@ -1,4 +1,4 @@
-// Copyright 2024 Google LLC
+// Copyright 2021-2024 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,3 +13,7 @@
 // limitations under the License.
 
 export * from './logging-utils';
+// 🌟 ADDED EXPORT FOR TYPES 🌟
+export type { LogFields, LogSeverity, AdhocDebugLogFunction, DebugLogBackend } from './types';
+
+
