+++ packages/logging-utils/src/logging-utils.ts
@@ -16,77 +16,14 @@ import {EventEmitter} from 'events';
 import * as process from 'process';
 import * as util from 'util';
 import {Colours} from './colours';
+import {LogFields, LogSeverity, AdhocDebugLogCallable, AdhocDebugLogFunction, DebugLogBackend} from './types';
 
 // Some functions (as noted) are based on the Node standard library, from
 // the following file:
 //
 // https://github.com/nodejs/node/blob/main/lib/internal/util/debuglog.js
 
-/**
- * This module defines an ad-hoc debug logger for Google Cloud Platform
- * client libraries in Node. An ad-hoc debug logger is a tool which lets
- * users use an external, unified interface (in this case, environment
- * variables) to determine what logging they want to see at runtime. This
- * isn't necessarily fed into the console, but is meant to be under the
- * control of the user. The kind of logging that will be produced by this
- * is more like "call retry happened", not "events you'd want to record
- * in Cloud Logger".
- *
- * More for Googlers implementing libraries with it:
- * go/cloud-client-logging-design
- */
-
-/**
- * Possible log levels. These are a subset of Cloud Observability levels.
- * https://cloud.google.com/logging/docs/reference/v2/rest/v2/LogEntry#LogSeverity
- */
-export enum LogSeverity {
-  DEFAULT = 'DEFAULT',
-  DEBUG = 'DEBUG',
-  INFO = 'INFO',
-  WARNING = 'WARNING',
-  ERROR = 'ERROR',
-}
-
-/**
- * A set of suggested log metadata fields.
- */
-export interface LogFields {
-  /**
-   * Log level - undefined/null === DEFAULT.
-   */
-  severity?: LogSeverity;
-
-  /**
-   * If this log is associated with an OpenTelemetry trace, you can put the
-   * trace ID here to pass on that association.
-   */
-  telemetryTraceId?: string;
-
-  /**
-   * If this log is associated with an OpenTelemetry trace, you can put the
-   * span ID here to pass on that association.
-   */
-  telemetrySpanId?: string;
-
-  /**
-   * This is a catch-all for any other items you might want to go into
-   * structured logs. Library implementers, please see the spec docs above
-   * for the items envisioned to go here.
-   */
-  other?: unknown;
-}
-
-/**
- * Adds typings for event sinks.
- */
-export declare interface AdhocDebugLogger {
-  on(
-    event: 'log',
-    listener: (fields: LogFields, args: unknown[]) => void,
-  ): this;
-  on(event: string, listener: Function): this;
-}
+// NOTE: The previous 'export declare interface AdhocDebugLogger' was correctly removed.
 
 /**
  * Our logger instance. This actually contains the meat of dealing
@@ -106,7 +43,7 @@ export class AdhocDebugLogger extends EventEmitter {
 
   /**
    * @param upstream The backend will pass a function that will be
-   *   called whenever our logger function is invoked.
+   * called whenever our logger function is invoked.
    */
   constructor(namespace: string, upstream: AdhocDebugLogCallable) {
     super();
@@ -162,66 +99,7 @@ export class AdhocDebugLogger extends EventEmitter {
  */
 export const placeholder = new AdhocDebugLogger('', () => {}).func;
 
-/**
- * When the user receives a log function (below), this will be the basic function
- * call interface for it.
- */
-export interface AdhocDebugLogCallable {
-  (fields: LogFields, ...args: unknown[]): void;
-}
-
-/**
- * Adds typing info for the EventEmitter we're adding to the returned function.
- *
- * Note that this interface may change at any time, as we're reserving the
- * right to add new backends at the logger level.
- *
- * @private
- * @internal
- */
-export interface AdhocDebugLogFunction extends AdhocDebugLogCallable {
-  instance: AdhocDebugLogger;
-  on(
-    event: 'log',
-    listener: (fields: LogFields, args: unknown[]) => void,
-  ): this;
-
-  debug(...args: unknown[]): void;
-  info(...args: unknown[]): void;
-  warn(...args: unknown[]): void;
-  error(...args: unknown[]): void;
-  sublog(namespace: string): AdhocDebugLogFunction;
-}
-
-/**
- * One of these can be passed to support a third-party backend, like "debug".
- * We're splitting this out because ESM can complicate optional module loading.
- *
- * Note that this interface may change at any time, as we're reserving the
- * right to add new backends at the logger level.
- *
- * @private
- * @internal
- */
-export interface DebugLogBackend {
-  /**
-   * Outputs a log to this backend.
-   *
-   * @param namespace The "system" that will be used for filtering. This may also
-   *   include a "subsystem" in the form "system:subsystem".
-   * @param fields Logging fields to be included as metadata.
-   * @param args Any parameters to passed to a utils.format() type formatter.
-   */
-  log(namespace: string, fields: LogFields, ...args: unknown[]): void;
-
-  /**
-   * Passes in the system/subsystem filters from the global environment variables.
-   * This lets the backend merge with any native ones.
-   *
-   * @param filters A list of wildcards matching systems or system:subsystem pairs.
-   */
-  setFilters(filters: string[]): void;
-}
+// 🌟 REMOVED: AdhocDebugLogCallable, AdhocDebugLogFunction, DebugLogBackend declarations
 
 /**
  * The base class for debug logging backends. It's possible to use this, but the
