+++ packages/logging-utils/tsconfig.json
@@ -6,7 +6,8 @@
     "outDir": "build",
     "noImplicitAny": true,
     "resolveJsonModule": true,
-    "esModuleInterop": true                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables `allowSyntheticDefaultImports` for type compatibility. */
+    "esModuleInterop": true,
+    "composite": true
   },
   "include": [
     "src/*.ts",
