+++ packages/nodejs-googleapis-common/package.json
@@ -39,8 +39,9 @@
   "license": "Apache-2.0",
   "dependencies": {
     "extend": "^3.0.2",
-    "gaxios": "7.1.3",
-    "google-auth-library": "10.5.0",
+    "gaxios": "file:../gaxios",
+    "google-auth-library": "file:../google-auth-library-nodejs",
+    "google-logging-utils": "file:../logging-utils",
     "qs": "^6.7.0",
     "url-template": "^2.0.8"
   },
@@ -57,7 +58,7 @@
     "@types/tmp": "^0.2.6",
     "@types/url-template": "^2.0.28",
     "c8": "^10.1.3",
-    "codecov": "^3.8.3",
+    "codecov": "^3.6.2",
     "gts": "^6.0.2",
     "http2spy": "^2.0.0",
     "is-docker": "^3.0.0",
